requirements.txt
duckduckgo-search==6.3.5
requests==2.32.3
Pillow==10.4.0
Jinja2==3.1.4
rich==13.9.4
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
